+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All changes to this project will be documented in this file. This version
 format adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [x.x.x] - xxxx-xx-xx
+
+- feat: Add support for default values in ci pipeline.
+
 ## [0.4.0] - 2025-11-09
 
 - feat: Add support for !reference tags.